Re factor policy representations and remove commented code
