+++ Protos/src/thinclab/tests/TestStructures.java
@@ -31,13 +31,13 @@ import thinclab.exceptions.ZeroProbabilityObsException;
 import thinclab.legacy.DD;
 import thinclab.legacy.OP;
 import thinclab.parsers.IPOMDPParser;
-import thinclab.policy.DynamicBeliefTree;
-import thinclab.policy.MJ;
-import thinclab.policy.PolicyGraph;
-import thinclab.policy.WalkablePolicyTree;
-import thinclab.policy.StaticBeliefTree;
-import thinclab.policy.ConditionalPlanGraph;
-import thinclab.policy.ConditionalPlanTree;
+import thinclab.representations.ConditionalPlanGraph;
+import thinclab.representations.ConditionalPlanTree;
+import thinclab.representations.DynamicBeliefTree;
+import thinclab.representations.MJ;
+import thinclab.representations.PolicyGraph;
+import thinclab.representations.StaticBeliefTree;
+import thinclab.representations.WalkablePolicyTree;
 import thinclab.solvers.OfflinePBVISolver;
 import thinclab.solvers.OfflineSymbolicPerseus;
 import thinclab.solvers.OnlineIPBVISolver;
